nearest_cars/requirements.txt
asgiref==3.8.0
Django==5.0.3
djangorestframework==3.15.0
flake8==7.0.0
flake8-broken-line==1.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
geographiclib==2.0
geopy==2.4.1
isort==5.13.2
mccabe==0.7.0
pep8-naming==0.13.3
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
